### 2024-07-26 #7038 @kthyng
@JohnVolk paper comments:
* need spaces before your references
* Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.
* the paper is too long — I would recommend shortening the "Design and Features" section and moving any information removed from there to the docs if it isn't already there.
