### 2024-07-27 25a3314
paper updates, shorten, reference fixes
