+++ paper/paper.md
@@ -39,19 +39,19 @@ This package has the functionality to download point data from a variety of grid
 
 # Design and Features
 
-``gridwxcomp`` is a Python 3 package that consists of five core submodules and two utility submodules (\autoref{fig:fig1}). Currently, ``gridwxcomp`` can process the following meteorological variables: air temperature (minimum and maximum), dew point temperature, shortwave radiation, wind speed, vapor pressure, relative humidity (minimum, maximum, and average), and grass (short) and alfalfa (tall) reference evapotranspiration (ET). The following gridded weather datasets are examples of those that are hosted on Google Earth Engine and available to be compared to station data using this package: CONUS404[@Rasmussen2023], ERA5-Land[@MuozSabater2021], gridMET[@Abatzoglou2013], NLDAS[@Mitchell2004], RTMA[@DePondeca2011], and spatial CIMIS[@Hart2009]. Other gridded weather datasets hosted on Google Earth can be accessed and compared against station data using ``gridwxcomp`` as long as the user has access to the dataset collection. 
+``gridwxcomp`` is a Python 3 package that consists of five core submodules and two utility submodules (\autoref{fig:fig1}). Currently, ``gridwxcomp`` can process the following meteorological variables: air temperature (minimum and maximum), dew point temperature, shortwave radiation, wind speed, vapor pressure, relative humidity (minimum, maximum, and average), and grass (short) and alfalfa (tall) reference evapotranspiration (ET). The following gridded weather datasets are examples of those that are hosted on Google Earth Engine and available to be compared to station data using this package: CONUS404 [@Rasmussen2023], ERA5-Land [@MuozSabater2021], gridMET [@Abatzoglou2013], NLDAS [@Mitchell2004], RTMA [@DePondeca2011], and spatial CIMIS [@Hart2009]. Other gridded weather datasets hosted on Google Earth can be accessed and compared against station data using ``gridwxcomp`` as long as the user has access to the dataset collection. 
 
 ![Flowchart diagram of submodules and data processing pipeline of ``gridwxcomp``.\label{fig:fig1}](figure1.pdf)
 
-Starting with ``prep_metadata``, this module parses metadata regarding one or more meteorological stations’ daily weather data. Regarding the gridded dataset to compare with, this script outputs a CSV text file with coordinates used for pairing each meteorological station with the gridded data. 
+The ``prep_metadata`` submodule parses metadata of meteorological stations. It also reprojects station coordiantes and outputs a reformatted CSV file. 
 
-The output file from ``prep_metadata`` is used by the next submodule ``ee_download`` which queries the Google Earth Engine for the previously specified weather data variables from the specified gridded product. It then downloads those data at gridcell centroids that are nearest to the corresponding station coordinates for the period that matches the station data. This process also updates the output file from ``prep_metadata`` to include the file paths of the downloaded gridded timeseries data. 
+The output file from ``prep_metadata`` is used by the ``ee_download`` submodule which queries the Google Earth Engine for the gridded weather data specifed in the configuration file. It downloads time series data at the corresponding station coordinates. 
 
[…]
