+++ paper/paper.bib
@@ -92,7 +92,7 @@ note = {also available at \url{https://www.mesonet.org/images/site/ASCE_Evapotra
   pages = {4349–4383}
 }
 @article{Hart2009,
-  title = {Daily reference evapotranspiration for California using satellite imagery and weather station measurement interpolation},
+  title = {Daily reference evapotranspiration for {C}alifornia using satellite imagery and weather station measurement interpolation},
   volume = {26},
   ISSN = {1029-0249},
   url = {http://dx.doi.org/10.1080/10286600802003500},
@@ -120,7 +120,7 @@ note = {also available at \url{https://www.mesonet.org/images/site/ASCE_Evapotra
   pages = {593–612}
 }
 @article{Mitchell2004,
-  title = {The multi‐institution North American Land Data Assimilation System (NLDAS): Utilizing multiple GCIP products and partners in a continental distributed hydrological modeling system},
+  title = {The multi‐institution {N}orth {A}merican {L}and {D}ata {A}ssimilation {S}ystem (NLDAS): Utilizing multiple GCIP products and partners in a continental distributed hydrological modeling system},
   volume = {109},
   ISSN = {0148-0227},
   url = {http://dx.doi.org/10.1029/2003JD003823},
