### 2024-07-27 f874694
paper clean up and shorten
