+++ paper/paper.md
@@ -33,23 +33,21 @@ Gridded weather data has become increasingly accessible and accurate over the re
 
 # Statement of Need
 
-Commonly, in situ measurements of weather data are used to validate and assess the bias and uncertainty in their gridded counterparts. Point biases can be interpolated to investigate spatial biases given sufficient density of measurement stations. Finally, maps of spatial bias can subsequently be used to adjust or correct the gridded weather data for the observed bias. ``gridwxcomp`` was developed precisely to streamline these objectives in a reproducible Python framework. 
+Commonly, in situ measurements of weather data are used to validate and assess the bias and uncertainty in their gridded counterparts. Point biases can be interpolated to investigate spatial biases given sufficient density of measurement stations. Maps of spatial bias can subsequently be used to adjust the gridded weather data for the observed bias. ``gridwxcomp`` was developed to streamline these objectives in a reproducible Python framework. 
 
 This package has the functionality to download point data from a variety of gridded meteorological datasets that are hosted on [Google Earth Engine](https://developers.google.com/earth-engine/datasets/) (e.g., NLDAS, ERA5, gridMET) and pair those with station data. It also has functionality to make comparison plots, calculate monthly bias ratios and metrics, and interpolate those data to make spatially complete georeferenced raster images of bias between the gridded and station data using multiple interpolation techniques such as inverse distance weighting and linear interpolation. As far as the authors know, this is the only open-source software that accomplishes these tasks.
 
 # Design and Features
 
-``gridwxcomp`` is a Python 3 package that consists of five core submodules and two utility submodules (\autoref{fig:fig1}). Currently, ``gridwxcomp`` can process the following meteorological variables: air temperature (minimum and maximum), dew point temperature, shortwave radiation, wind speed, vapor pressure, relative humidity (minimum, maximum, and average), and grass (short) and alfalfa (tall) reference evapotranspiration (ET). The following gridded weather datasets are examples of those that are hosted on Google Earth Engine and available to be compared to station data using this package: CONUS404 [@Rasmussen2023], ERA5-Land [@MuozSabater2021], gridMET [@Abatzoglou2013], NLDAS [@Mitchell2004], RTMA [@DePondeca2011], and spatial CIMIS [@Hart2009]. Other gridded weather datasets hosted on Google Earth can be accessed and compared against station data using ``gridwxcomp`` as long as the user has access to the dataset collection. 
+``gridwxcomp`` is a Python 3 package that consists of five core submodules and two utility submodules (\autoref{fig:fig1}). ``gridwxcomp`` can process the following meteorological variables: air temperature (minimum and maximum), dew point temperature, shortwave radiation, wind speed, vapor pressure, relative humidity (minimum, maximum, and average), and grass (short) and alfalfa (tall) reference evapotranspiration (ET). Daily gridded weather datasets hosted on Google Earth Engine can be accessed and compared against station data using ``gridwxcomp`` as long as the user has access to the dataset collection. Eexample public datasets include: CONUS404 [@Rasmussen2023], ERA5-Land [@MuozSabater2021], gridMET [@Abatzoglou2013], NLDAS [@Mitchell2004], RTMA [@DePondeca2011], and spatial CIMIS [@Hart2009]. 
 
 ![Flowchart diagram of submodules and data processing pipeline of ``gridwxcomp``.\label{fig:fig1}](figure1.pdf)
 
-The ``prep_metadata`` submodule parses metadata of meteorological stations. It also reprojects station coordiantes and outputs a reformatted CSV file. 
+The ``prep_metadata`` submodule parses metadata of meteorological stations, including reprojection of station coordinates. The output file from ``prep_metadata`` is used by the ``ee_download`` submodule which queries Google Earth Engine for the gridded weather data specifed by the user and downloads time series data at the corresponding station coordinates. 
 
[…]
