### 2024-09-03 #7038 @ArcticSnow
Hello @hugoledoux , I'd be happy reviewing this library. It looks interesting. I will need 2 or 3 weeks to find some time though to focus on the review.
