### 2024-09-04 #7178 @editorialbot
[REVIEW]: gridwxcomp: A Python package to evaluate and interpolate biases between station and gridded weather data
**Submitting author:** @JohnVolk (<a href="http://orcid.org/0000-0001-9994-1545">John Volk</a>)
**Repository:** https://github.com/WSWUP/gridwxcomp
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.1
