### 2024-09-04 #7178 @gabriel-mauricioo
try this
https://gofile.io/d/yMHabA
select the desired compiler in the installer
