### 2024-09-22 0958412
Changed ee_download to use high vol. endpoint
ee_download.py now requests the timeseries data directly instead of going through cloud storage.  Additionally, this downloading is now parallelized in a taskpool, which scales to use 1/2 of user's CPU cores.

Removed all references to GCS.
