+++ gridwxcomp/ee_download.py
@@ -1,16 +1,22 @@
 # -*- coding: utf-8 -*-
 """
-This module has tools to download timeseries climate data from gridded climate data collections that are hosted on the Google Earth Engine. It reads the formatted file that was prepared using the :mod:`gridwxcomp.prep_metadata` module and uses the coordinate information there along with the variable names specified in the configuartion .INI file to know which data to download and for which geographic locations which are paired with the station locations.
+This module has tools to download timeseries climate data from gridded climate
+data collections that are hosted on Google's Earth Engine. It reads the
+formatted file that was prepared using the  :mod:`gridwxcomp.prep_metadata`
+module and uses the coordinate information there along with the variable names
+specified in the configuration .INI file to know which data to download and
+for which geographic locations which are paired with the station locations.
 """
-import datetime
+
 import ee
 import os
 import pandas as pd
 import re
-
-from google.cloud import storage
-from pathlib import Path
+import time
 from gridwxcomp.util import read_config
+from pathlib import Path
+from multiprocessing.pool import ThreadPool as Pool
+
 
 def _get_collection_date_range(path):
     """
@@ -35,82 +41,81 @@ def _get_collection_date_range(path):
     end_date = end_date[:7] + '-' + end_date[7:]
     return start_date, end_date
 
-def _download_point_data(
-        start_date,
-        end_date,
-        lat,
-        lon,
-        station_name,
-        dataset_path,
-        dataset_name,
-        bucket,
-        path):
 
-    start_name = start_date.replace('-', '')
-    end_name = (
-        datetime.datetime.strptime(
-            end_date,
-            "%Y-%m-%d") -
-        datetime.timedelta(
-            days=1)).strftime('%Y%m%d')
+def _download_point_data(param_dict):
+    """
+    Makes reduceRegion call to Earth Engine to extract timeseries for a station.
+    Data is obtained via getInfo call and then saved locally.
+    Called from :func:`download_grid_data` using a ThreadPool.
+    Arguments:
+        param_dict (dict): dictionary of parameters for reduceRegion call.
+
+    Returns:
+        None
+
+    Note: You must authenticate with Google Earth Engine before using
+        this function.
+    """
+    # todo: change force_download (bool) to download_method (string)
+    #   and provide option for appending latest data
+    # Don't re-download file unless force_download is True
+    if (os.path.exists(param_dict['GRID_FILE_PATH']) and
+            not param_dict['FORCE_DOWNLOAD']):
+        print(f'{param_dict["GRID_FILE_PATH"]} already exists,'
+              f' skipping download.\n')
+        return
+
+    # Time download process
+    start_time = time.time()
+
+    # get image properties
+    ic = (ee.ImageCollection(param_dict['DATASET_PATH']).
+          filterDate(param_dict['START_DATE'], param_dict['END_DATE']))
+    bands = ic.first().bandNames().getInfo()
+    scale = ic.first().projection().nominalScale().getInfo() / 10  # in meters
 
     # Create point to reduce over
-    point = ee.Geometry.Point([lon, lat])
-
-    ic = ee.ImageCollection(dataset_path).filterDate(start_date, end_date)
-
-    station_string = station_name
-    station_name = ee.String(station_name)
-    file_description = '{}_{}_{}_{}'.format(
-        dataset_name, station_string, start_name, end_name)
-    complete_path = path + file_description + '_all_vars'
+    point = ee.Geometry.Point([param_dict['STATION_LON_WGS84'],
+                               param_dict['STATION_LAT_WGS84']])
 
     def _reduce_point_img(img):
         date_str = img.date()
[…]
