+++ demo.py
@@ -1,6 +1,5 @@
 import ee
-import pyproj
-from gridwxcomp.prep_metadata import prep_metadata, get_subgrid_bounds
+from gridwxcomp.prep_metadata import prep_metadata
 from gridwxcomp.ee_download import download_grid_data
 from gridwxcomp.plot import daily_comparison, monthly_comparison, station_bar_plot
 from gridwxcomp.calc_bias_ratios import calc_bias_ratios
@@ -8,19 +7,17 @@ from gridwxcomp.spatial import make_grid, interpolate
 from gridwxcomp.util import reproject_crs_for_bounds
 
 
-# initialize earth engine
-ee.Initialize()
 # name of the dataset comparison is being made with
-gridded_dataset_name = 'conus404'
-# Path to station metadata file with lat/long coords
-station_meta_path = './gridwxcomp/example_data/Station_Data.txt'
+gridded_dataset_name = 'era5land'
 # local path for config file
-config_path = f'./gridwxcomp/example_data/gridwxcomp_config_{gridded_dataset_name}.ini'
+config_path = f'gridwxcomp_config_{gridded_dataset_name}.ini'
 
-# name of bucket data will be exported to once the earth engine calls complete
-export_bucket = 'openet'
-# path in bucket to export to
-export_path = f'bias_correction_gridwxcomp_testing/gridwxcomp_{gridded_dataset_name}/'
+# initialize earth engine
+ee.Authenticate()
+ee.Initialize(opt_url='https://earthengine-highvolume.googleapis.com')
+
+# Path to station metadata file with lat/long coords
+station_meta_path = 'openet_station_data/conus_comparison_station_list.csv'
 
 # local path for prep_metadata output
 gridwxcomp_input = f'{gridded_dataset_name}_gridwxcomp_metadata.csv'
@@ -79,33 +76,27 @@ projection_dict = {
         'crs_id': 'ESRI:102004'}
 }
 
-# User has option to automatically pull bounds, aee TODO in function over assumptions
+# User has option to automatically pull bounds, see TODO in function over assumptions
 # projection_dict['wgs84']['bounds'] = get_subgrid_bounds(gridwxcomp_input, config_path, buffer=25)
 
 projection_dict['lcc']['bounds'] = reproject_crs_for_bounds(
     projection_dict['wgs84']['bounds'], projection_dict['lcc']['resolution'],
     projection_dict['wgs84']['crs_id'], projection_dict['lcc']['crs_id'], 0)
 
-
 '''
 download_grid_data
-   The purpose of this module is to make calls to the earth engine API to export gridded timeseries at the station
-        points contained within the station metadata file generated by prep_metadata. Exported data will be saved to
-        a cloud storage bucket that the user specifies.
+    The purpose of this module is to make calls to the earth engine API to 
+        export timeseries for the gridded dataset at the station locations 
+        contained within the station metadata file generated by prep_metadata.
+    
+    The data is saved locally as it is downloaded, and the metadata file
+        generated by prep_metadata is amended with paths to the downloaded files.
 
-   The user will need to manually download the data via gsutil before proceeding further. The module amends the
-        station metadata file to have paths for the gridded dataset as they would exist locally once the user has
-        downloaded them.
-
-   Example: the example run is with the dataset conus404, the gridded path will be filled in such that if you
-        download the 'conus404' folder from the bucket to the local directory of trial_runs.py it will work without
-        further modification.
 '''
 download_grid_data(
     gridwxcomp_input,
     config_path,
-    export_bucket=export_bucket,
-    export_path=export_path,
+    local_folder=None,
[…]
