### 2024-10-07 86ac3f9
add co-author roles and fix typo joss paper
