+++ paper/paper.md
@@ -39,7 +39,7 @@ This package has the functionality to download point data from a variety of grid
 
 # Design and Features
 
-``gridwxcomp`` is a Python 3 package that consists of five core submodules and two utility submodules (\autoref{fig:fig1}). ``gridwxcomp`` can process the following meteorological variables: air temperature (minimum and maximum), dew point temperature, shortwave radiation, wind speed, vapor pressure, relative humidity (minimum, maximum, and average), and grass (short) and alfalfa (tall) reference evapotranspiration (ET). Daily gridded weather datasets hosted on Google Earth Engine can be accessed and compared against station data using ``gridwxcomp`` as long as the user has access to the dataset collection. Eexample public datasets include: CONUS404 [@Rasmussen2023], ERA5-Land [@MuozSabater2021], gridMET [@Abatzoglou2013], NLDAS [@Mitchell2004], RTMA [@DePondeca2011], and spatial CIMIS [@Hart2009]. 
+``gridwxcomp`` is a Python 3 package that consists of five core submodules and two utility submodules (\autoref{fig:fig1}). ``gridwxcomp`` can process the following meteorological variables: air temperature (minimum and maximum), dew point temperature, shortwave radiation, wind speed, vapor pressure, relative humidity (minimum, maximum, and average), and grass (short) and alfalfa (tall) reference evapotranspiration (ET). Daily gridded weather datasets hosted on Google Earth Engine can be accessed and compared against station data using ``gridwxcomp`` as long as the user has access to the dataset collection. Example public datasets include: CONUS404 [@Rasmussen2023], ERA5-Land [@MuozSabater2021], gridMET [@Abatzoglou2013], NLDAS [@Mitchell2004], RTMA [@DePondeca2011], and spatial CIMIS [@Hart2009]. 
 
 ![Flowchart diagram of submodules and data processing pipeline of ``gridwxcomp``.\label{fig:fig1}](figure1.pdf)
 
@@ -54,6 +54,17 @@ The ``plot`` submodule in ``gridwxcomp`` provides tools for generating interacti
 # Research Enabled by gridwxcomp
 
 The most significant application of ``gridwxcomp`` was the development of bias correction surfaces that are applied to gridded reference evapotransipiration (ETo) data which are key inputs to some of the remote sensing ET models that comprise the OpenET platform [@Volk2024;@Melton2021]. Daily data from approximately 800 weather stations located in irrigated agricultural sites were curated, and the American Society of Civil Engineers (ASCE) standardized Penman-Monteith reference ET equation [@allen2005] was used to estimate ETo at the stations. Then ``gridwxcomp`` was used to pair these data with the nearest ETo data from the gridMET [@Abatzoglou2013] dataset over temporally consistent periods. The long-term average monthly ratios for station ETo relative to the gridded ETo were calculated for each point and saved as georeferenced data by ``gridwxcomp`` and were subsequently spatially interpolated using a kriging approach. The interpolated monthly surfaces are used within the OpenET platform to correct gridMET ETo data before it is used by most of the remote sensing ET models as a major scaling flux. 
+
+# Co-author Roles
+
+``gridwxcomp`` was developed through the following efforts:
+
+* John M. Volk: Conceptualization, Software, Validation, Writing & Editting
+* Christian Dunkerly: Conceptualization, Software, Validation
+* Christopher Pearson: Conceptualization, Software, Validation
+* Charles Morton: Conceptualization, Software
+* Justin L. Huntington: Conceptualization, Funding acquisition, Resources 
+
 
 # Acknowledgments
 
