### 2024-10-22 #7178 @ArcticSnow
Hi,
As I started on the review, I am trying to get `gridwxcomp` to run on my local machine. I was able to install it without issues. I am however getting lost in the Google Earth Engine process. I have `ee.Authenticate()` returning `True`. I created a project on the google portal, but struggle to get the `ee` API to Initialize `ee.Initialize(project='testgridwxcomp')`.

Any clue what may be the issue? I have never used `ee` and I am not sure if as a public person I can access the data.
