### 2024-10-22 #7178 @JohnVolk
Hi Simon,

When initializing EE there is no data being accessed, later the code will pull from public Google EE data assets so that should not be an issue. Can you post the error you are seeing from `ee.Initialize() `?
