### 2024-10-23 #7178 @ArcticSnow
Thank you John for your quick support!
Interestingly, when running as you suggested `ee.Initialize()` without specifying `project`, it prompt me a useful error message. I was able to register the project, and clear this step. Google Earth engine from the Python API is new to me.
