### 2024-11-21 WSWUP/gridwxcomp#28 @dvalters
How to contribute/community guidelines
Hi, as part of the JOSS review at https://github.com/openjournals/joss-reviews/issues/7178#issuecomment-2491778558, could you add a contributing section to readme or link to a section in the docs if I've missed this? Thanks
