### 2024-11-22 c830839
add contributing and guidelines, addressing #28
