+++ docs/source/contribute.rst
@@ -1,0 +1,50 @@
+How to Contribute to `gridwxcomp`
+=================================
+
+We welcome community involvement and contributions aimed to improve `gridwxcomp`. This includes spotting and fixing bugs, adding new features, or improving documentation, all contributions are much appreciated. Please follow the guidelines below to ensure that your contributions are well-received and incorporated efficiently.
+
+Reporting Issues
+----------------
+If you encounter a bug or unexpected behavior, please report it by creating an issue on the repository's issue tracker. When submitting a bug report, include the following:  
+
+* A clear description of the issue  
+* Steps to reproduce the issue  
+* Expected vs. actual results  
+* Any error messages or logs  
+* Details of your environment (e.g., operating system, software version)  
+
+Contributing Code
+-----------------
+
+**Note**: If you would like to propose a major change or add significant features to gridwxcomp, please reach out to the developers before you begin. Start by creating an issue with a "new feature" tag in the repository. This allows us to give feedback and guidance during the development process, and will keep your work aligned with the project's goals.
+
+1. **Fork the Repository**: Start by forking the `gridwxcomp` repository to your GitHub account. This will allow you to work on your own copy of the project.
+   
+2. **Create a Branch**: Always create a new branch for your changes. Avoid making changes directly on the `master` branch. Branches should be named descriptively based on the feature or bug you're addressing (e.g., `fix-typo-in-docs`, `add-rainfall-forecast`).
+
+3. **Write Tests**: If your contribution involves a new feature or a bug fix, please add corresponding tests using ``pytest``. 
+   
+4. **Follow Code Style Guidelines**: Please adhere to the PEP 8 style guide used in the project. 
+
+5. **Commit Messages**: When committing your changes, make sure your commit messages are clear, concise, and follow a logical structure. A good format is:
+   ```
+   Type of change: Short description
+   More detailed explanation of what the change does and why it is needed.
+   ```
+
+6. **Pull Request**: Once you have completed your changes, create a pull request from your branch to the `master` branch of the ``gridwxcomp`` repository. Please ensure your pull request description explains the purpose of your changes and any relevant context. 
+
+Contributing to Documentation
+-----------------------------
+Documentation is just as important as code. If you spot an area where the documentation can be improved or updated, feel free to contribute! 
+
+* **Readability**: Ensure the documentation is clear and easy to understand. If you encounter unclear sections, feel free to suggest improvements.
+* **Formatting**: Please use reStructuredText syntax for formatting. 
+
+Community Guidelines
+--------------------
+We strive to maintain a welcoming and respectful community. Please follow these guidelines:
+
+* **Be respectful**: Treat other contributors with respect. Everyone is here to improve the software.
+* **Be constructive**: Offer feedback that is helpful and aimed at improving the project.
+
