### 2024-11-22 d0f258f
minor font style change in contributing docs section
