### 2024-11-26 #7178 @dvalters
I'm also struggling to get past the Earth Engine Initialization point using Google Cloud projects. I've tried running `ee.Initialize()` and `ee.Initialize(project="ee-myproject") but I am getting errors such as:

```
 project, credentials)
    268     project = self._determine_default(project)
    270 if project is None:
--> 271     raise EnvironmentError(
    272         "Project was not passed and could not be "
    273         "determined from the environment."
    274     )
    276 if isinstance(project, bytes):
    277     project = project.decode("utf-8")

OSError: Project was not passed and could not be determined from the environment.
```

I have tried passing the project name but I still get the same error
