### 2024-11-26 #7178 @dvalters
My running feeling when reviewing this submission so far is that the software seems very tightly coupled with Google Earth Engine, and now Google Cloud Projects, which has made it time-consuming/difficult to access the data required for testing purposes. That's not a criticism of the software per se which looks to have a very useful application. But my feeling is that it falls short of the following JOSS criteria:

> Data sharing: If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.

Is there a way to provide a standalone, offline, test data set? Or could the tutorial documentation be improved to guide people not familiar with setting up Google Earth Engine and Google Cloud projects?

Thanks for your patience so far
