### 2024-11-26 #7178 @JohnVolk
Hi @dvalters, I'm working on testing with the new project ID GEE requirement and will add what I find here and to the docs, if for some reason it cannot be resolved, I can add a standalone dataset for you to continue testing. Thanks for your update. The reason we went with GEE as a tool for accessing data is simply because it hosts so many different weather and climate datasets publicly. I'm sure we can get this straightened out soon and improve the docs with more details to prevent this from happening for others.
