### 2024-11-26 #7178 @JohnVolk
A couple things to check, @dvalters:

* Can you check on the Google Cloud Console that your project-ID is currently enabled and being used, https://console.cloud.google.com/

* Try running `ee.Authenticate()` before `ee.initialize()`, this should help you setup the GEE project the first time.

If neither of those work also try adding the following if you are running Jupyter Notebook (based off the error you showed):

`import os
os.environ["GOOGLE_CLOUD_PROJECT"] = "your-project-id"
`
You could also try adding your project ID to your ~/.bashrc file using the following line:

`export GOOGLE_CLOUD_PROJECT="your-project-id"`

Please update with how it goes.
