### 2024-12-03 b5f42f4
Add GEE setup instructions to docs
