### 2024-12-03 #7178 @JohnVolk
@dvalters I'm just following up on your GEE account set up, have you been able to get it working? Also, to improve the software and help other user's from experiencing similar problems I added a section [Getting Started with Google Earth Engine](https://gridwxcomp.readthedocs.io/en/latest/gee.html) that should cover in detail how to set up an account and get it working properly before continuing with the online tutorial. Thanks for bringing this up, it will be a big help for many I'm sure!
