### 2024-12-19 c886342
re-added precipitation after it was removed on previous build
