+++ gridwxcomp/example_data/gridwxcomp_config_agera5.ini
@@ -67,6 +67,7 @@ station_rhmin_col = RHMin (%)
 station_rhavg_col = RHAvg (%)
 station_eto_col = ETo (mm)
 station_etr_col = ETr (mm)
+station_prcp_col = Precip (mm)
 
 gridded_date_col = date
 gridded_tmax_col = Temperature_Air_2m_Max_24h
@@ -80,6 +81,7 @@ gridded_rhmin_col =
 gridded_rhavg_col = Relative_Humidity_2m_15h
 gridded_eto_col = 
 gridded_etr_col = 
+gridded_prcp_col = Precipitation_Flux
 
 [UNITS]
 # For the parameters in this section, enter the corresponding units from the options commented above.
@@ -107,3 +109,7 @@ gridded_rh_units = percent
 # inches, mm
 station_et_units = mm
 gridded_et_units = mm
+
+# inches, mm
+station_prcp_units = mm
+gridded_prcp_units = mm
