### 2024-12-19 1e68f53
adjusted tests to work with ee high endpoint approach
