### 2024-12-19 1a911b6
bump version, docs update
