### 2025-01-02 #7178 @JohnVolk
Hi @ArcticSnow, I hope you have had a nice holiday. I know it's a busy time of the year but I would kindly inquire if you have made progress on the review of the software and paper? Best regards!
