### 2025-01-06 #7178 @ArcticSnow
Hello @JohnVolk ,

I am glad to tell you that I was able to run through all steps of the tutorial as in the documentation. While I understand the rational to use Google Earth Engine API and therefore make the toolbox fitted to this cloud platform, the methods developed within `gridwxcomp` would be of interest to many if available to agnostic sources of data (zarr or netcdf based). This may be choices of development for the future of this library.
As is, the toolbox is operational and openly accessible despite its full dependency on GEE (itself not open and requiring billing information). The documentation is well written in accordance to the code library. Usefulness of the toolbox is well justified in the manuscript.

If the dependency on GEE is not a problem within the scope of JOSS, I happily support this publication to move forward.

I apologize for the very long delay in my response and wish you a happy new year 2025!
Simon Filhol

Minor comments:

line 12, 14, 17: change "in situ" to "in-situ" in italic
line 17-21: This sentence is very long and could be split to convey the message in simpler terms.
line 42: typo in "Eexample"

Problem with GEE export bucket and path. Require billing information, unlike specified in documentation tutorial. Station csv files were then created in the GEE bucket, and not downloaded on my local machine.
