### 2025-01-06 #7178 @JohnVolk
Hi @ArcticSnow,

Thank you for the review, and I am glad to hear that you feel that the software is useful. The minor review points regarding the paper text will be corrected. With regard to the GEE export- in the last update and version bump we changed the software to avoid saving gridded data to a Google bucket before downloading the files locally. Now in version 0.2.1, the files should directly downloaded locally, skipping the buckets. And, there should be no need for billing information for setting up GEE, but thanks for the feedback, hopefully the newly added documentation on setting up GEE helps others avoid these issues.

Happy 2025 to you as well!
John
