### 2025-01-06 964c712
minor paper and docs edits
