+++ paper/paper.md
@@ -29,11 +29,11 @@ bibliography: paper.bib
 
 # Introduction
 
-Gridded weather data has become increasingly accessible and accurate over the recent decades, and such data enables a variety of applications and research that require spatially continuous and high spatio-temporal resolution data [@Thornton2021;@Rasmussen2023;@MuozSabater2021]. Gridded weather data are often developed from an assimilation of data production and measurement techniques, including the incorporation of in situ observational data networks, land surface modeling techniques, and remote sensing techniques. Although well-curated in situ measurements of weather variables are typically more accurate than gridded products, they provide data at a single point in space, and they involve difficulties and expenses related to deployment and sensor calibration and maintenance, resulting in incomplete spatial and temporal coverage. Gaps in spatial and temporal coverage in in situ weather data are often filled by gridded data products, however, the increased coverage provided by gridded data comes with the tradeoff of increased uncertainty and potential for bias [@Blankenau2020] that is introduced in the modeling and statistical data assimilation techniques used for gridded data production that can be difficult to characterize and quantify. 
+Gridded weather data has become increasingly accessible and accurate over the recent decades, and such data enables a variety of applications and research that require spatially continuous and high spatio-temporal resolution data [@Thornton2021;@Rasmussen2023;@MuozSabater2021]. Gridded weather data are often developed from an assimilation of data production and measurement techniques, including the incorporation of *in-situ* observational data networks, land surface modeling techniques, and remote sensing techniques. Although well-curated *in-situ* measurements of weather variables are typically more accurate than gridded products, they provide data at a single point in space, and they involve difficulties and expenses related to deployment and sensor calibration and maintenance, resulting in incomplete spatial and temporal coverage. Gaps in spatial and temporal coverage in *in-situ* weather data are often filled by gridded data products. The increased coverage provided by data filling in gridded datasets comes with the tradeoff of uncertainty and potential for bias [@Blankenau2020] that are introduced by gridded data development and assimilation techniques. 
 
 # Statement of Need
 
-Commonly, in situ measurements of weather data are used to validate and assess the bias and uncertainty in their gridded counterparts. Point biases can be interpolated to investigate spatial biases given sufficient density of measurement stations. Maps of spatial bias can subsequently be used to adjust the gridded weather data for the observed bias. ``gridwxcomp`` was developed to streamline these objectives in a reproducible Python framework. 
+Commonly, *in-situ* measurements of weather data are used to validate and assess the bias and uncertainty in their gridded counterparts. Point biases can be interpolated to investigate spatial biases given sufficient density of measurement stations. Maps of spatial bias can subsequently be used to adjust the gridded weather data for the observed bias. ``gridwxcomp`` was developed to streamline these objectives in a reproducible Python framework. 
 
 This package has the functionality to download point data from a variety of gridded meteorological datasets that are hosted on [Google Earth Engine](https://developers.google.com/earth-engine/datasets/) (e.g., NLDAS, ERA5, gridMET) and pair those with station data. It also has functionality to make comparison plots, calculate monthly bias ratios and metrics, and interpolate those data to make spatially complete georeferenced raster images of bias between the gridded and station data using multiple interpolation techniques such as inverse distance weighting and linear interpolation. As far as the authors know, this is the only open-source software that accomplishes these tasks.
 
