### 2025-01-09 #7178 @hugoledoux
okay, I am back from holidays, great that all finished their reviews 🙌

I have proof-read the paper and please fix those typos:
- fix the missing DOIs (see above)
- your affiliation should also contain the city and the country
- l.8: I think that "data" is plural but you put "has" and "enables". I know that opinions differ on this, just pointing it out... I think putting "datasets" makes it simpler/clearer. You decide.
- l.35: "Python 3" Since Python 2.7 was officially deprecated >5y ago I would skip the 3, it's obvious to me
- l.47: specied => specified
- l.67: evapotransipiration => evapotranspiration
- the section "Co-Author Roles" is not necessary, but you can leave it.
- if you do leave it, then at l.81: Editting => Editing
- l.134: north america => North America

after this is done, could you:
