+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @ARTICLE{Thornton2021,
-  title     = "Gridded daily weather data for North America with comprehensive
+  title     = "Gridded daily weather data for {N}orth {A}merica with comprehensive
                uncertainty quantification",
   author    = "Thornton, Peter E and Shrestha, Rupesh and Thornton, Michele and
                Kao, Shih-Chieh and Wei, Yaxing and Wilson, Bruce E",
@@ -31,6 +31,7 @@
   pages     = "190",
   month     =  jul,
   year      =  2021,
+  doi = {10.1038/s41597-021-00973-0},
   copyright = "https://creativecommons.org/licenses/by/4.0",
   language  = "en"
 }
@@ -56,6 +57,7 @@
   number    =  106376,
   pages     = "106376",
   month     =  dec,
+  doi       = {10.1016/j.agwat.2020.106376},
   year      =  2020,
   language  = "en"
 }
