+++ paper/paper.md
@@ -21,7 +21,7 @@ authors:
   - name: Justin L. Huntington
     affiliation: 1 
 affiliations:
-  - name: Desert Research Institute
+  - name: Desert Research Institute, Reno, USA
     index: 1
 date: April 2024
 bibliography: paper.bib
@@ -29,7 +29,7 @@ bibliography: paper.bib
 
 # Introduction
 
-Gridded weather data has become increasingly accessible and accurate over the recent decades, and such data enables a variety of applications and research that require spatially continuous and high spatio-temporal resolution data [@Thornton2021;@Rasmussen2023;@MuozSabater2021]. Gridded weather data are often developed from an assimilation of data production and measurement techniques, including the incorporation of *in-situ* observational data networks, land surface modeling techniques, and remote sensing techniques. Although well-curated *in-situ* measurements of weather variables are typically more accurate than gridded products, they provide data at a single point in space, and they involve difficulties and expenses related to deployment and sensor calibration and maintenance, resulting in incomplete spatial and temporal coverage. Gaps in spatial and temporal coverage in *in-situ* weather data are often filled by gridded data products. The increased coverage provided by data filling in gridded datasets comes with the tradeoff of uncertainty and potential for bias [@Blankenau2020] that are introduced by gridded data development and assimilation techniques. 
+Gridded weather data have become increasingly accessible and accurate over the recent decades, and such data enable a variety of applications and research that require spatially continuous and high spatio-temporal resolution data [@Thornton2021;@Rasmussen2023;@MuozSabater2021]. Gridded weather data are often developed from an assimilation of data production and measurement techniques, including the incorporation of *in-situ* observational data networks, land surface modeling techniques, and remote sensing techniques. Although well-curated *in-situ* measurements of weather variables are typically more accurate than gridded products, they provide data at a single point in space, and they involve difficulties and expenses related to deployment and sensor calibration and maintenance, resulting in incomplete spatial and temporal coverage. Gaps in spatial and temporal coverage in *in-situ* weather data are often filled by gridded data products. The increased coverage provided by data filling in gridded datasets comes with the tradeoff of uncertainty and potential for bias [@Blankenau2020] that are introduced by gridded data development and assimilation techniques. 
 
 # Statement of Need
 
@@ -39,11 +39,11 @@ This package has the functionality to download point data from a variety of grid
 
 # Design and Features
 
-``gridwxcomp`` is a Python 3 package that consists of five core submodules and two utility submodules (\autoref{fig:fig1}). ``gridwxcomp`` can process the following meteorological variables: air temperature (minimum and maximum), dew point temperature, shortwave radiation, wind speed, vapor pressure, relative humidity (minimum, maximum, and average), and grass (short) and alfalfa (tall) reference evapotranspiration (ET). Daily gridded weather datasets hosted on Google Earth Engine can be accessed and compared against station data using ``gridwxcomp`` as long as the user has access to the dataset collection. Example public datasets include: CONUS404 [@Rasmussen2023], ERA5-Land [@MuozSabater2021], gridMET [@Abatzoglou2013], NLDAS [@Mitchell2004], RTMA [@DePondeca2011], and spatial CIMIS [@Hart2009]. 
[…]
