### 2025-01-09 #7178 @JohnVolk
Hi @hugoledoux Thanks for your quick reply! I have completed the paper edits and data release as requested, here is the info:

The version tag of the archived release is: [v0.2.1](https://github.com/WSWUP/gridwxcomp/tree/v0.2.1)

The Zenodo DOI for the archived version is: 10.5281/zenodo.14624093
