### 2025-01-10 #7178 @hugoledoux
@JohnVolk the names on the paper and on zenodo are not the same, and we require that they are... Middles are in the paper but not on zenodo.

You can edit either, but they have to be the same.
