### 2025-01-10 #7178 @JohnVolk
@hugoledoux sounds good, the Zenodo metadata has been updated, thanks!
