### 2025-01-13 #7178 @hugoledoux
Thank you @ArcticSnow @dvalters for your reviews, I hope you learned something in the process!

@JohnVolk an EiC will now process and finalise your paper, it's possible they ask for some small changes. This should be done in the coming ~2days.
