### 2025-01-13 #7178 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
