### 2025-01-13 #7178 @ArcticSnow
Congratulations @JohnVolk for the publication. And thank you @JohnVolk  and @hugoledoux for your patience. I am fond of the JOSS reviewing format and process. Working myself with meteorological/climate observations and models, having a new toolbox available working towards facilitating assessment of models and observation will be usefull to the community, no doubt.
