### 2025-01-13 #7178 @JohnVolk
Thank you @ArcticSnow, @dvalters , and @hugoledoux for your time and effort reviewing and editing! @ArcticSnow I'm glad to hear that you may be able to use these tools in your work. We hope that this publication will make the software better known and ultimately facilitate climate/modeling analyses and research applications in the broader community. Have a great week.
