### 2025-01-14 #7178 @kthyng
@JohnVolk Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example, "united states" is not capitalized but please check for others.
