### 2025-01-14 4caeb7e
fix paper citation capitalization
