+++ paper/paper.bib
@@ -49,7 +49,7 @@
 }
 @ARTICLE{Blankenau2020,
   title     = "An evaluation of gridded weather data sets for the purpose of
-               estimating reference evapotranspiration in the United States",
+               estimating reference evapotranspiration in the {U}nited {S}tates",
   author    = "Blankenau, Philip A and Kilic, Ayse and Allen, Richard",
   journal   = "Agric. Water Manag.",
   publisher = "Elsevier BV",
@@ -67,20 +67,20 @@
    issn = {1093-474X},
    journal = {JAWRA Journal of the American Water Resources Association},
    month = {11},
-   title = {OpenET: Filling a Critical Data Gap in Water Management for the Western United States},
+   title = {OpenET: {F}illing a critical data gap in water management for the western {U}nited {S}tates},
    url = {https://onlinelibrary.wiley.com/doi/10.1111/1752-1688.12956},
    year = {2021},
 }
 @book{allen2005,
 author = {Richard G. Allen  and Ivan A. Walter  and Ronald L. Elliott  and Terry A. Howell  and Daniel Itenfisu  and Marvin E. Jensen  and Richard L. Snyder },
-title = {The ASCE Standardized Reference Evapotranspiration Equation},
+title = {The ASCE {S}tandardized {R}eference {E}vapotranspiration {E}quation},
 publisher = {American Society of Civil Engineers},
 year = {2005},
 DOI = {10.1061/9780784408056},
 note = {also available at \url{https://www.mesonet.org/images/site/ASCE_Evapotranspiration_Formula.pdf}},
 }
 @article{MuozSabater2021,
-  title = {ERA5-Land: a state-of-the-art global reanalysis dataset for land applications},
+  title = {ERA5-{L}and: a state-of-the-art global reanalysis dataset for land applications},
   volume = {13},
   ISSN = {1866-3516},
   url = {http://dx.doi.org/10.5194/essd-13-4349-2021},
